Main/USBIdDatabase.h: Signed/unsigned conversion issues. bugref:9790
svn:sync-xref-src-repo-rev: r139235

+++ src/VBox/Main/include/USBIdDatabase.h
@@ -1,4 +1,4 @@
-/* $Id: USBIdDatabase.h 82968 2020-02-04 10:35:17Z knut.osmundsen@oracle.com $ */
+/* $Id: USBIdDatabase.h 85258 2020-07-11 23:55:41Z knut.osmundsen@oracle.com $ */
 /** @file
  * USB device vendor and product ID database.
  */
@@ -91,7 +91,7 @@ public:
     {
         char szTmp[USB_ID_DATABASE_MAX_STRING * 2];
         ssize_t cchTmp = RTBldProgStrTabQueryString(&s_StrTab, pStr->off, pStr->cch, szTmp, sizeof(szTmp));
-        return RTCString(szTmp, RT_MAX(cchTmp, 0));
+        return RTCString(szTmp, (size_t)RT_MAX(cchTmp, 0));
     }
 
 private:
